+++ src/index.js
@@ -7,6 +7,7 @@ export const rules = {
   'export': require('./rules/export'),
   'no-mutable-exports': require('./rules/no-mutable-exports'),
   'extensions': require('./rules/extensions'),
+  'no-restricted-paths': require('./rules/no-restricted-paths'),
 
   'no-named-as-default': require('./rules/no-named-as-default'),
   'no-named-as-default-member': require('./rules/no-named-as-default-member'),